+++ common/src/main/java/foundry/veil/api/resource/type/TextResource.java
@@ -1,16 +1,24 @@
 package foundry.veil.api.resource.type;
 
+import foundry.veil.Veil;
 import foundry.veil.api.resource.VeilResourceAction;
 import foundry.veil.api.resource.VeilResourceInfo;
 import foundry.veil.api.resource.VeilResourceManager;
 import foundry.veil.impl.resource.action.TextEditAction;
+import imgui.extension.texteditor.TextEditorLanguage;
 import org.jetbrains.annotations.ApiStatus;
+import org.jetbrains.annotations.Nullable;
 
 import java.io.IOException;
 import java.util.List;
+import java.util.function.Supplier;
 
 @ApiStatus.Internal
-public record TextResource(VeilResourceInfo resourceInfo, Type type) implements VeilTextResource<TextResource> {
+public record TextResource(
+        VeilResourceInfo resourceInfo,
+        Type type,
+        @Nullable TextEditorLanguage languageDefinition
+) implements VeilTextResource<TextResource> {
 
     @Override
     public List<VeilResourceAction<TextResource>> getActions() {
@@ -33,14 +41,20 @@ public record TextResource(VeilResourceInfo resourceInfo, Type type) implements 
 
     public enum Type {
         TEXT(".txt", 0xED0F),
-        JSON(".json", 0xECCD);
+        JSON(".json", 0xECCD, TextEditorLanguage::Json);
 
         private final String extension;
         private final int icon;
+        private final Supplier<TextEditorLanguage> languageDefinition;
 
-        Type(String extension, int icon) {
+        Type(String extension, int icon, @Nullable Supplier<TextEditorLanguage> languageDefinition) {
             this.extension = extension;
             this.icon = icon;
+            this.languageDefinition = languageDefinition;
+        }
+
+        Type(String extension, int icon) {
+            this(extension, icon, null);
         }
 
         public String getExtension() {
@@ -49,6 +63,10 @@ public record TextResource(VeilResourceInfo resourceInfo, Type type) implements 
 
         public int getIcon() {
             return this.icon;
+        }
+
+        public @Nullable TextEditorLanguage getLanguageDefinition() {
+            return Veil.IMGUIMC && this.languageDefinition != null ? this.languageDefinition.get() : null;
         }
     }
 }