+++ common/src/main/java/foundry/veil/mixin/shader_recompile/accessor/ShaderRecompileProgramAccessor.java
@@ -1,11 +1,11 @@
-package foundry.veil.mixin.dynamicbuffer.accessor;
+package foundry.veil.mixin.shader_recompile.accessor;
 
 import com.mojang.blaze3d.shaders.Program;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Accessor;
 
 @Mixin(Program.class)
-public interface DynamicBufferProgramAccessor {
+public interface ShaderRecompileProgramAccessor {
 
     @Accessor
     int getId();