+++ changelog.md
@@ -1,3 +1,5 @@
-- Add mutable getters for light data
-- Implement a particle editor for Quasar ([#163](https://github.com/FoundryMC/Veil/pull/163))
-- Improve trails module ([#164](https://github.com/FoundryMC/Veil/pull/164))
+- Update to ImGuiMC 2.0.0 ([#161](https://github.com/FoundryMC/Veil/pull/161))
+- Add spotlight ([#167](https://github.com/FoundryMC/Veil/pull/167))
+- Add raymarched in-scattering as an option for deferred lights ([#168](https://github.com/FoundryMC/Veil/pull/168))
+- Add DDA occlusion for directional lights ([#171](https://github.com/FoundryMC/Veil/pull/171))
+